refactor!: merge json path with the existing url storage and clarify usage of backend types
