+++ htsget-config/src/config/advanced/mod.rs
@@ -17,10 +17,6 @@ use std::path::PathBuf;
 pub mod auth;
 pub mod callout;
 pub mod cors;
-#[cfg(feature = "url")]
-pub mod http;
-#[cfg(feature = "url")]
-pub mod json_path;
 pub mod regex_location;
 #[cfg(feature = "url")]
 pub mod url;
